src/func_tests/test_requirements.txt
git-python==1.0.*
requests==2.*
numpy==1.24.*
